renamed the main package to be com.revature.hydra.gateway
