+++ src/main/java/com/revature/hydra/gateway/ZuulServiceApplication.java
@@ -1,15 +1,12 @@
-package com.revature.caliber;
+package com.revature.hydra.gateway;
 
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
 import org.springframework.cloud.netflix.zuul.EnableZuulProxy;
 import org.springframework.context.annotation.Bean;
 import org.springframework.web.bind.annotation.CrossOrigin;
-import org.springframework.web.cors.CorsConfiguration;
-import org.springframework.web.cors.UrlBasedCorsConfigurationSource;
-import org.springframework.web.filter.CorsFilter;
 
-import com.revature.caliber.filters.RedirectToAuthenticationPreFilter;
+import com.revature.hydra.gateway.filters.RedirectToAuthenticationPreFilter;
 
 import springfox.documentation.builders.PathSelectors;
 import springfox.documentation.builders.RequestHandlerSelectors;
