--- src/main/resources/application.properties
@@ -1,1 +1,0 @@
-security.basic.enabled=false
