+++ src/main/resources/application.yml
@@ -1,0 +1,64 @@
+server:
+  port: 8081
+
+spring:
+  application:
+    name: zuul
+
+eureka:
+  client:
+    serviceUrl:
+      defaultZone: http://localhost:8761/eureka/
+
+zuul:
+  routes:
+    frontend:
+      path: /frontend/**
+      serviceId: frontend
+    grade:
+      path: /grade/**
+      serviceId: grade
+    pdf:
+      path: /pdf/**
+      serviceId: pdf
+    types:
+      path: /types/**
+      serviceId: types
+    trainer:
+      path: /trainer/**
+      serviceId: trainer
+    assessment:
+      path: /assessment/**
+      serviceId: assessment
+    trainee:
+      path: /trainee/**
+      serviceId: trainee
+    batch:
+      path: /batch/**
+      serviceId: batch
+    salesforce:
+      path: /salesforce/**
+      serviceId: salesforce
+    note:
+      path: /note/**
+      serviceId: note
+    address:
+      path: /address/**
+      serviceId: address
+    category:
+      path: /category/**
+      serviceId: category
+    panel:
+      path: /panel/**
+      serviceId: panel
+    panelfeedback:
+      path: /panelfeedback/**
+      serviceId: panelfeedback
+    security:
+      path: /security/**
+      serviceId: security
+    reporting:
+      path: /reporting/**
+      serviceId: reporting
+ribbon: 
+  ReadTimeout: 300000
