+++ src/main/resources/bootstrap.yml
@@ -3,7 +3,7 @@ spring:
     name: zuul
   cloud: 
     config: 
-      uri: http://localhost:8888
+      uri: http://localhost:9000
 
 security:
   basic:
